Refactor engine namespaces
