--- src/protagonist/Engine/Ingest/Completion/IEngineAssetRepository.cs
@@ -1,20 +1,0 @@
-﻿using DLCS.Core.Types;
-using DLCS.Model.Assets;
-
-namespace Engine.Ingest.Completion;
-
-public interface IEngineAssetRepository
-{
-    /// <summary>
-    /// Update database with ingested asset.
-    /// </summary>
-    /// <param name="asset">Asset to update</param>
-    /// <param name="imageLocation">ImageLocation, optional as may have exited prior to creation</param>
-    /// <param name="imageStorage">ImageStorage, optional as may have exited prior to creation</param>
-    /// <param name="cancellationToken">Current cancellation token</param>
-    /// <returns>True if successful</returns>
-    Task<bool> UpdateIngestedAsset(Asset asset, ImageLocation? imageLocation, ImageStorage? imageStorage,
-        CancellationToken cancellationToken = default);
-
-    ValueTask<Asset?> GetAsset(AssetId assetId, CancellationToken cancellationToken = default);
-}
