+++ src/protagonist/Engine/Ingest/Timebased/Transcode/IMediaTranscoder.cs
@@ -1,0 +1,9 @@
+﻿namespace Engine.Ingest.Timebased.Transcode;
+
+public interface IMediaTranscoder
+{
+    /// <summary>
+    /// Initiate a request to start transcoding asset.
+    /// </summary>
+    Task<bool> InitiateTranscodeOperation(IngestionContext context, CancellationToken token = default);
+}
