--- src/protagonist/Engine/Ingest/Workers/IAssetIngestorWorker.cs
@@ -1,18 +1,0 @@
-﻿using DLCS.Model.Customers;
-using DLCS.Model.Messaging;
-
-namespace Engine.Ingest.Workers;
-
-/// <summary>
-/// Interface for operations related to ingesting assets.
-/// Different workers will handle different types of asset
-/// </summary>
-public interface IAssetIngesterWorker
-{
-    /// <summary>
-    /// Ingest provided asset using given CustomerOriginStrategy
-    /// </summary>
-    Task<IngestResult> Ingest(IngestAssetRequest ingestAssetRequest,
-        CustomerOriginStrategy customerOriginStrategy,
-        CancellationToken cancellationToken = default);
-}
