--- src/protagonist/Engine/Ingest/Workers/Persistence/AssetMoverBase.cs
@@ -1,24 +1,0 @@
-﻿using DLCS.Core.Types;
-using DLCS.Model.Storage;
-
-namespace Engine.Ingest.Workers.Persistence;
-
-/// <summary>
-/// Base class with helpers for moving assets to alternative storage
-/// </summary>
-public abstract class AssetMoverBase
-{
-    protected readonly IStorageRepository StorageRepository;
-
-    protected AssetMoverBase(IStorageRepository storageRepository)
-    {
-        StorageRepository = storageRepository;
-    }
-    
-    protected async Task<bool> VerifyFileSize(AssetId assetId, long size)
-    {
-        var customerHasEnoughSize = await StorageRepository.VerifyStoragePolicyBySize(assetId.Customer,
-            size);
-        return customerHasEnoughSize;
-    }
-}
